chore: refactored code so functions live in the correct place
