+++ mcp-server/src/core/direct-functions/show-task.js
@@ -109,12 +109,20 @@ export async function showTaskDirect(args, log) {
  * @param {string} args.id - The Jira issue key to show details for.
  * @param {boolean} [args.withSubtasks=false] - If true, will fetch subtasks for the parent task
  * @param {boolean} [args.includeImages=true] - If true, will fetch and include image attachments
+ * @param {boolean} [args.includeContext=false] - If true, will include related tickets and PR context
+ * @param {number} [args.maxRelatedTickets=10] - Maximum number of related tickets to fetch in context
  * @param {Object} log - Logger object
  * @returns {Promise<Object>} - Task details result { success: boolean, data?: any, error?: { code: string, message: string } }
  */
 export async function showJiraTaskDirect(args, log) {
 	// Destructure expected args
-	const { id, includeImages = true } = args;
+	const {
+		id,
+		includeImages = true,
+		includeContext = false,
+		withSubtasks = false,
+		maxRelatedTickets = 10
+	} = args;
 
 	// Validate task ID
 	const taskId = id;
@@ -140,9 +148,9 @@ export async function showJiraTaskDirect(args, log) {
 		// Use the dedicated function from jira-utils.js to fetch task details
 		const jiraTaskResult = await fetchJiraTaskDetails(
 			taskId,
-			args.withSubtasks,
+			withSubtasks,
 			log,
-			{ includeImages }
+			{ includeImages, includeContext, maxRelatedTickets }
 		);
 
 		// Restore normal logging before returning
